+++ lib/handlers/va_handler.dart
@@ -3,6 +3,8 @@ import 'package:botika_va/models/webhook_model.dart';
 
 abstract class BotikaVaHandler {
   void onVaError(String message);
-  void onVaResponse(MessageModel msg, List<DownloadVideoModel> videos);
-  void onVaResponseVoice(MessageModel msg, List<String?> audios);
+  void onVaResponse(
+      String responseId, MessageModel msg, List<DownloadVideoModel> videos);
+  void onVaResponseVoice(
+      String responseId, MessageModel msg, List<String?> audios);
 }