+++ lib/models/webhook_model.dart
@@ -145,6 +145,17 @@ class MessageModel {
     return data;
   }
 
+  List<ButtonModel> getButtons() {
+    List<ButtonModel> returns = [];
+    if (type == "button") {
+      List<dynamic> buttons = value as List<dynamic>;
+      for (dynamic b in buttons) {
+        returns.add(ButtonModel.fromJson(b));
+      }
+    }
+    return returns;
+  }
+
   List<String> getChunk() {
     if (type != "text") {
       return [];
@@ -182,5 +193,27 @@ class MessageModel {
       }
     }
     return false;
+  }
+}
+
+class ButtonModel {
+  String? type;
+  String? text;
+  String? payload;
+
+  ButtonModel({this.type, this.text, this.payload});
+
+  ButtonModel.fromJson(Map<String, dynamic> json) {
+    type = json['type'];
+    text = json['text'];
+    payload = json['payload'];
+  }
+
+  Map<String, dynamic> toJson() {
+    Map<String, dynamic> data = <String, dynamic>{};
+    data['type'] = type;
+    data['text'] = text;
+    data['payload'] = payload;
+    return data;
   }
 }